MBL-14682: Moved announcement tests to AnnouncementInteractionTest
